refactor: fix overlap modal [#132]
